# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  email: string;
  name?: string | null;
}

export interface SignupRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  name?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  user: AuthUser;
}

export interface UpdateProfileRequest {
  name: string;
}

export interface ChangePasswordRequest {
  currentPassword: string;
  /** @minLength 8 */
  newPassword: string;
}

export interface ErrorResponse {
  error: string;
}

export interface DeleteAccountRequest {
  password: string;
}

export interface PracticeSessionDto {
  /**
     * ISO date, YYYY-MM-DD
     * @pattern ^\d{4}-\d{2}-\d{2}$
     */
  date: string;
  /**
     * Seconds practiced that day, capped at one full day
     * @minimum 0
     * @maximum 86400
     */
  duration: number;
  /**
     * Swings analyzed that day
     * @minimum 0
     * @maximum 100000
     */
  swings?: number;
}

export type SwingRecordDtoOrigin = typeof SwingRecordDtoOrigin[keyof typeof SwingRecordDtoOrigin];


export const SwingRecordDtoOrigin = {
  mine: 'mine',
  pro: 'pro',
} as const;

export type SwingRecordDtoGameMode = typeof SwingRecordDtoGameMode[keyof typeof SwingRecordDtoGameMode];


export const SwingRecordDtoGameMode = {
  long: 'long',
  short: 'short',
} as const;

export type SwingRecordDtoClub = typeof SwingRecordDtoClub[keyof typeof SwingRecordDtoClub] | null;


export const SwingRecordDtoClub = {
  tee: 'tee',
  approach: 'approach',
  shortgame: 'shortgame',
  putting: 'putting',
} as const;

export interface SwingRecordDto {
  /**
     * Client-generated id, unique per account (not globally)
     * @minLength 1
     * @maxLength 64
     */
  id: string;
  /**
     * ISO date, YYYY-MM-DD
     * @pattern ^\d{4}-\d{2}-\d{2}$
     */
  date: string;
  /**
     * Epoch milliseconds. Bounded to the range JS Date can represent.
     * @minimum 0
     * @maximum 8640000000000000
     */
  timestamp: number;
  /**
     * @minLength 1
     * @maxLength 64
     */
  swingId: string;
  origin: SwingRecordDtoOrigin;
  /** @maxLength 120 */
  golferName: string;
  gameMode: SwingRecordDtoGameMode;
  club: SwingRecordDtoClub;
  /**
     * @minimum 0
     * @maximum 100
     */
  ratio: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  accuracy: number;
}

export interface AccountExport {
  /** ISO 8601 timestamp of when this export was generated */
  exportedAt: string;
  user: AuthUser;
  sessions: PracticeSessionDto[];
  swingRecords: SwingRecordDto[];
}

export interface SyncPayload {
  /** @maxItems 3650 */
  sessions: PracticeSessionDto[];
  /** @maxItems 500 */
  swingRecords: SwingRecordDto[];
}

export interface SyncRequest {
  /** @maxItems 3650 */
  sessions: PracticeSessionDto[];
  /** @maxItems 500 */
  swingRecords: SwingRecordDto[];
  /**
     * Ids the user deleted locally. The server removes these from the account before merging, so a deletion propagates instead of the record being resurrected by the next pull.
     * @maxItems 500
     * @items.minLength 1
     * @items.maxLength 64
     */
  deletedSwingRecordIds?: string[];
}

export type ShotCategory = typeof ShotCategory[keyof typeof ShotCategory];


export const ShotCategory = {
  tee: 'tee',
  approach: 'approach',
  shortgame: 'shortgame',
  putting: 'putting',
} as const;

export interface TempoVideoDto {
  id: string;
  category: ShotCategory;
  name: string;
  event: string;
  year: number;
  club: string;
  ratio: number;
  duration: number;
  backswing: number;
  downswing: number;
  result?: string | null;
  /** YouTube video id (not the full URL). Null until a clip has been sourced/uploaded for this entry. */
  youtubeId?: string | null;
  clipStartSec?: number | null;
  clipEndSec?: number | null;
  sortOrder: number;
}

export interface TempoVideoInput {
  category: ShotCategory;
  name: string;
  event: string;
  year: number;
  club: string;
  ratio: number;
  duration: number;
  backswing: number;
  downswing: number;
  result?: string;
  youtubeId?: string;
  clipStartSec?: number;
  clipEndSec?: number;
  sortOrder?: number;
}

export interface TempoVideoPatch {
  category?: ShotCategory;
  name?: string;
  event?: string;
  year?: number;
  club?: string;
  ratio?: number;
  duration?: number;
  backswing?: number;
  downswing?: number;
  result?: string;
  youtubeId?: string;
  clipStartSec?: number;
  clipEndSec?: number;
  sortOrder?: number;
}

export type ListTempoVideosParams = {
category?: ShotCategory;
};