# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Account creation and authentication
  - name: sync
    description: Local-first sync of practice sessions and swing history
  - name: tempoVideos
    description: Reference-pro tempo entries shown on the Tempos tab
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input, or an account with that email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid email or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProfile
      tags: [auth]
      summary: Update the current user's display name
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAccount
      tags: [auth]
      summary: Permanently delete the current account and all of its data
      description: >-
        Requires the account password to be re-entered. Deleting the user row
        cascades to practice_sessions and swing_records, so nothing is left
        behind. Irreversible — the client should confirm before calling.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAccountRequest"
      responses:
        "204":
          description: Account deleted
        "400":
          description: Password is incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me/export:
    get:
      operationId: exportAccountData
      tags: [auth]
      summary: Export everything stored server-side for the current account
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Full account export
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountExport"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/password:
    patch:
      operationId: changePassword
      tags: [auth]
      description: >-
        Changing the password revokes every token issued before it (the
        account's token version is bumped), so the response carries a freshly
        signed token for the calling device. Clients must persist it or they
        will be signed out on their next request.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed, with a replacement token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Current password is incorrect, or new password is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /sync:
    post:
      operationId: sync
      tags: [sync]
      summary: Push local practice sessions/swing records, get back the merged authoritative set
      description: >-
        Local-first sync: the client sends its current local snapshot, the
        server upserts it (sessions merged by max duration/swings per day,
        swing records deduped by id) and returns the full merged dataset for
        this account. The client replaces its local storage with the
        response.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncRequest"
      responses:
        "200":
          description: Merged sync state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncPayload"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tempo-videos:
    get:
      operationId: listTempoVideos
      tags: [tempoVideos]
      summary: List reference-pro tempo entries, optionally filtered by category
      parameters:
        - name: category
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ShotCategory"
      responses:
        "200":
          description: Tempo video entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TempoVideoDto"
    post:
      operationId: createTempoVideo
      tags: [tempoVideos]
      summary: Add a tempo video entry (admin only)
      security:
        - adminAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TempoVideoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TempoVideoDto"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tempo-videos/{id}:
    patch:
      operationId: updateTempoVideo
      tags: [tempoVideos]
      summary: Update a tempo video entry, e.g. attach a youtubeId once sourced (admin only)
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TempoVideoPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TempoVideoDto"
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No entry with that id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTempoVideo
      tags: [tempoVideos]
      summary: Remove a tempo video entry (admin only)
      security:
        - adminAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted
        "401":
          description: Not authenticated as admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminAuth:
      type: apiKey
      in: header
      name: x-admin-token
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AuthUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
          nullable: true
      required:
        - id
        - email
    SignupRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        name:
          type: string
      required:
        - email
        - password
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AuthUser"
      required:
        - token
        - user
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
      required:
        - currentPassword
        - newPassword
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    DeleteAccountRequest:
      type: object
      properties:
        password:
          type: string
      required:
        - password
    AccountExport:
      type: object
      properties:
        exportedAt:
          type: string
          description: ISO 8601 timestamp of when this export was generated
        user:
          $ref: "#/components/schemas/AuthUser"
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/PracticeSessionDto"
        swingRecords:
          type: array
          items:
            $ref: "#/components/schemas/SwingRecordDto"
      required:
        - exportedAt
        - user
        - sessions
        - swingRecords
    # Bounds below are deliberate, not decorative: an unbounded `duration`
    # combined with the server's greatest(local, server) merge would let one
    # bad client permanently poison a day's practice total with no way to
    # correct it. See docs/AUDIT.md F3/S7.
    PracticeSessionDto:
      type: object
      properties:
        date:
          type: string
          description: ISO date, YYYY-MM-DD
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        duration:
          type: integer
          description: Seconds practiced that day, capped at one full day
          minimum: 0
          maximum: 86400
        swings:
          type: integer
          description: Swings analyzed that day
          minimum: 0
          maximum: 100000
      required:
        - date
        - duration
    SwingRecordDto:
      type: object
      properties:
        id:
          type: string
          description: Client-generated id, unique per account (not globally)
          minLength: 1
          maxLength: 64
        date:
          type: string
          description: ISO date, YYYY-MM-DD
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        timestamp:
          type: integer
          format: int64
          description: Epoch milliseconds. Bounded to the range JS Date can represent.
          minimum: 0
          maximum: 8640000000000000
        swingId:
          type: string
          minLength: 1
          maxLength: 64
        origin:
          type: string
          enum: [mine, pro]
        golferName:
          type: string
          maxLength: 120
        gameMode:
          type: string
          enum: [long, short]
        club:
          type: string
          nullable: true
          enum: [tee, approach, shortgame, putting, null]
        ratio:
          type: number
          minimum: 0
          maximum: 100
        accuracy:
          type: integer
          minimum: 0
          maximum: 100
      required:
        - id
        - date
        - timestamp
        - swingId
        - origin
        - golferName
        - gameMode
        - club
        - ratio
        - accuracy
    SyncPayload:
      type: object
      properties:
        sessions:
          type: array
          maxItems: 3650
          items:
            $ref: "#/components/schemas/PracticeSessionDto"
        swingRecords:
          type: array
          maxItems: 500
          items:
            $ref: "#/components/schemas/SwingRecordDto"
      required:
        - sessions
        - swingRecords
    SyncRequest:
      type: object
      properties:
        sessions:
          type: array
          maxItems: 3650
          items:
            $ref: "#/components/schemas/PracticeSessionDto"
        swingRecords:
          type: array
          maxItems: 500
          items:
            $ref: "#/components/schemas/SwingRecordDto"
        deletedSwingRecordIds:
          type: array
          maxItems: 500
          description: >-
            Ids the user deleted locally. The server removes these from the
            account before merging, so a deletion propagates instead of the
            record being resurrected by the next pull.
          items:
            type: string
            minLength: 1
            maxLength: 64
      required:
        - sessions
        - swingRecords
    ShotCategory:
      type: string
      enum: [tee, approach, shortgame, putting]
    TempoVideoDto:
      type: object
      properties:
        id:
          type: string
        category:
          $ref: "#/components/schemas/ShotCategory"
        name:
          type: string
        event:
          type: string
        year:
          type: integer
        club:
          type: string
        ratio:
          type: number
        duration:
          type: number
        backswing:
          type: number
        downswing:
          type: number
        result:
          type: string
          nullable: true
        youtubeId:
          type: string
          nullable: true
          description: YouTube video id (not the full URL). Null until a clip has been sourced/uploaded for this entry.
        clipStartSec:
          type: number
          nullable: true
        clipEndSec:
          type: number
          nullable: true
        sortOrder:
          type: integer
      required:
        - id
        - category
        - name
        - event
        - year
        - club
        - ratio
        - duration
        - backswing
        - downswing
        - sortOrder
    TempoVideoInput:
      type: object
      properties:
        category:
          $ref: "#/components/schemas/ShotCategory"
        name:
          type: string
        event:
          type: string
        year:
          type: integer
        club:
          type: string
        ratio:
          type: number
        duration:
          type: number
        backswing:
          type: number
        downswing:
          type: number
        result:
          type: string
        youtubeId:
          type: string
        clipStartSec:
          type: number
        clipEndSec:
          type: number
        sortOrder:
          type: integer
      required:
        - category
        - name
        - event
        - year
        - club
        - ratio
        - duration
        - backswing
        - downswing
    TempoVideoPatch:
      type: object
      properties:
        category:
          $ref: "#/components/schemas/ShotCategory"
        name:
          type: string
        event:
          type: string
        year:
          type: integer
        club:
          type: string
        ratio:
          type: number
        duration:
          type: number
        backswing:
          type: number
        downswing:
          type: number
        result:
          type: string
        youtubeId:
          type: string
        clipStartSec:
          type: number
        clipEndSec:
          type: number
        sortOrder:
          type: integer