# lib/api-zod/src/generated/types/accountExport.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUser } from './authUser';
import type { PracticeSessionDto } from './practiceSessionDto';
import type { SwingRecordDto } from './swingRecordDto';

export interface AccountExport {
  /** ISO 8601 timestamp of when this export was generated */
  exportedAt: string;
  user: AuthUser;
  sessions: PracticeSessionDto[];
  swingRecords: SwingRecordDto[];
}