# lib/api-zod/src/generated/types/practiceSessionDto.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PracticeSessionDto {
  /**
     * ISO date, YYYY-MM-DD
     * @pattern ^\d{4}-\d{2}-\d{2}$
     */
  date: string;
  /**
     * Seconds practiced that day, capped at one full day
     * @minimum 0
     * @maximum 86400
     */
  duration: number;
  /**
     * Swings analyzed that day
     * @minimum 0
     * @maximum 100000
     */
  swings?: number;
}